Creation of 2 controllers. Added exception handling on delete

+++ src/main/java/com/bmdb/web/ActorController.java
@@ -3,7 +3,10 @@ package com.bmdb.web;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.bmdb.business.Actor;
 import com.bmdb.db.ActorRepo;
@@ -37,8 +40,25 @@ public class ActorController {
 	}
 
 	@DeleteMapping("/{id}")
-	public void delete(@PathVariable int id) {
-		actorRepo.deleteById(id);
+	public Optional<Actor> delete(@PathVariable int id) {
+		Optional<Actor> actor = actorRepo.findById(id);
+		if (actor.isPresent()) {
+			try {
+				actorRepo.deleteById(id);
+			} catch (DataIntegrityViolationException dive) {
+				// catch exception when movie exists as FK for another table
+				System.err.println(dive.getRootCause().getMessage());
+				throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR,
+						"Foreign Key Constraint Issue - Actor ID:" + id + " is referred to elsewhere.");
+			} catch (Exception e) {
+				e.printStackTrace();
+				throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR,
+						"Exception caught during actor delete");
+			}
+		} else {
+			System.err.println("Actor delete error - No actor found for ID:" + id);
+		}
+		return actor;
 	}
 	
 }
